refactor: ShellType -> Shell, drop venv and PowerShell support
