+++ src/tools/task_runner/mod.rs
@@ -1,7 +1,7 @@
 pub mod task_config;
 pub use task_config::TaskConfig;
 
-pub mod shell_type;
+pub mod shell;
 
 mod runner;
 pub use runner::TaskRunner;
